fix: keep ownershipTypeFromJson/periodTypeFromJson internal
Move to map_helpers.dart, which isn't re-exported, instead of dropping
the leading underscore in entitlement_info_wrapper.dart which leaked
them into the public API via object_wrappers.dart.
