+++ lib/models/map_helpers.dart
@@ -1,7 +1,36 @@
+import 'entitlement_info_wrapper.dart';
 import 'period_unit.dart';
 import 'product_category.dart';
 import 'store.dart';
 import 'verification_result.dart';
+
+OwnershipType ownershipTypeFromJson(dynamic value) {
+  switch (value) {
+    case 'PURCHASED':
+      return OwnershipType.purchased;
+    case 'FAMILY_SHARED':
+      return OwnershipType.familyShared;
+    case 'UNKNOWN':
+      return OwnershipType.unknown;
+    default:
+      return OwnershipType.unknown;
+  }
+}
+
+PeriodType periodTypeFromJson(dynamic value) {
+  switch (value) {
+    case 'INTRO':
+      return PeriodType.intro;
+    case 'NORMAL':
+      return PeriodType.normal;
+    case 'TRIAL':
+      return PeriodType.trial;
+    case 'PREPAID':
+      return PeriodType.prepaid;
+    default:
+      return PeriodType.unknown;
+  }
+}
 
 VerificationResult verificationResultFromJson(dynamic value) {
   switch (value) {
